Resolve a few miscellaneous warnings
git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@24703 dc483132-0cff-0310-8789-dd5450dbe970

+++ src/lib/crypto/krb/crypto_length.c
@@ -42,7 +42,7 @@ krb5_c_crypto_length(krb5_context context, krb5_enctype enctype,
         *size = 0;
         break;
     case KRB5_CRYPTO_TYPE_DATA:
-        *size = (size_t)~0; /* match Heimdal */
+        *size = (unsigned int)~0; /* match Heimdal */
         break;
     case KRB5_CRYPTO_TYPE_HEADER:
     case KRB5_CRYPTO_TYPE_PADDING:
